refactor: Split some sheet manipulation functions into smaller chunks
